remove AbstractStreamingMarshal class
